--- a/vllm/model_executor/layers/mamba/mamba2_metadata.py
+++ b/vllm/model_executor/layers/mamba/mamba2_metadata.py
@@ -13,7 +13,6 @@ from vllm.attention.backends.xformers import XFormersMetadata
 
 @dataclass
 class Mamba2Metadata:
-    has_prefill: bool
 
     has_initial_states: torch.Tensor
     prep_initial_states: bool
@@ -22,9 +21,6 @@ class Mamba2Metadata:
     seq_idx: torch.Tensor
     chunk_indices: torch.Tensor
     chunk_offsets: torch.Tensor
-
-    num_prefills: int
-    num_decodes: int
 
 
 def _seq_idx_to_chunk_indices_offsets(seq_idx, chunk_size: int):
@@ -63,7 +59,6 @@ def _seq_idx_to_chunk_indices_offsets(seq_idx, chunk_size: int):
 
 def prepare_mamba2_metadata(
     chunk_size: int,
-    # input_ids: torch.Tensor,
     attn_metadata: AttentionMetadata,
 ) -> Mamba2Metadata:
 
@@ -71,9 +66,6 @@ def prepare_mamba2_metadata(
     # NOTE: in V0 we assume prefills are before decodes
     num_prefills = attn_metadata.num_prefills
     num_prefill_tokens = attn_metadata.num_prefill_tokens
-    num_decodes = attn_metadata.num_decode_tokens
-
-    has_prefill = num_prefills > 0
 
     # Need flags to indicate if there are initial states
     # currently we really only support the FlashAttention backend
@@ -90,7 +82,7 @@ def prepare_mamba2_metadata(
     # Compute seq_idx, chunk_indices and chunk_offsets for prefill only
     seq_idx = None
     chunk_indices, chunk_offsets = None, None
-    if has_prefill:
+    if num_prefills > 0:
         query_start_loc = attn_metadata.query_start_loc[:num_prefills + 1]
         seq_idx = torch.repeat_interleave(torch.arange(
             num_prefills, dtype=torch.int32, device=query_start_loc.device),
@@ -105,12 +97,9 @@ def prepare_mamba2_metadata(
             chunk_indices, chunk_offsets = _seq_idx_to_chunk_indices_offsets(
                 seq_idx, chunk_size)
 
-    return Mamba2Metadata(has_prefill=has_prefill,
-                          has_initial_states=has_initial_states,
+    return Mamba2Metadata(has_initial_states=has_initial_states,
                           prep_initial_states=prep_initial_states,
                           chunk_size=chunk_size,
                           seq_idx=seq_idx,
                           chunk_indices=chunk_indices,
-                          chunk_offsets=chunk_offsets,
-                          num_prefills=num_prefills,
-                          num_decodes=num_decodes)
+                          chunk_offsets=chunk_offsets)