As a follow up to PR #16942 

The CUDA Causal Conv1d kernel has a similar problem to the mamba2 ssd prefill kernels. It doesn't perform well when chunked prefill is turned ON resulting in the input batch to have mixed prefill and decode requests. This PR splits the requests, and we observe a big total throughput improvement for benchmark_serving.py with SharedGPT v3 workload.

### This PR chunked prefill ON
```
vllm serve ibm-ai-platform/Bamba-9B --port 9999 
```

```
python benchmarks/benchmark_serving.py --model ibm-ai-platform/Bamba-9B  --dataset-name sharegpt     --dataset-path /net/storage149/mnt/md0/ccyang/github.com/ShareGPT_V3/ShareGPT_V3_unfiltered_cleaned_split.json --ignore-eos --port 9999 
============ Serving Benchmark Result ============
Successful requests:                     1000      
Benchmark duration (s):                  47.63     
Total input tokens:                      215201    
Total generated tokens:                  198343    
Request throughput (req/s):              21.00     
Output token throughput (tok/s):         4164.34   
Total Token throughput (tok/s):          8682.61   
---------------Time to First Token----------------
Mean TTFT (ms):                          15089.14  
Median TTFT (ms):                        13729.31  
P99 TTFT (ms):                           35227.37  
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          56.06     
Median TPOT (ms):                        53.62     
P99 TPOT (ms):                           116.86    
---------------Inter-token Latency----------------
Mean ITL (ms):                           48.57     
Median ITL (ms):                         48.07     
P99 ITL (ms):                            118.30    
==================================================
```
### This PR chunked prefill OFF

```
vllm serve ibm-ai-platform/Bamba-9B --port 9999 --no-enable-chunked-prefill --max_model_len=4096
```

```
============ Serving Benchmark Result ============
Successful requests:                     1000      
Benchmark duration (s):                  52.72     
Total input tokens:                      215201    
Total generated tokens:                  198343    
Request throughput (req/s):              18.97     
Output token throughput (tok/s):         3761.94   
Total Token throughput (tok/s):          7843.62   
---------------Time to First Token----------------
Mean TTFT (ms):                          16965.86  
Median TTFT (ms):                        15540.24  
P99 TTFT (ms):                           40363.40  
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          68.57     
Median TPOT (ms):                        63.19     
P99 TPOT (ms):                           259.66    
---------------Inter-token Latency----------------
Mean ITL (ms):                           56.02     
Median ITL (ms):                         58.01     
P99 ITL (ms):                            121.75    
==================================================
```


### Main (d9ac9e3) chunked prefill on
```
vllm serve ibm-ai-platform/Bamba-9B --port 9999 
```

```
python benchmarks/benchmark_serving.py --model ibm-ai-platform/Bamba-9B  --dataset-name sharegpt     --dataset-path /net/storage149/mnt/md0/ccyang/github.com/ShareGPT_V3/ShareGPT_V3_unfiltered_cleaned_split.json --ignore-eos --port 9999 
============ Serving Benchmark Result ============
Successful requests:                     1000      
Benchmark duration (s):                  183.82    
Total input tokens:                      215201    
Total generated tokens:                  198343    
Request throughput (req/s):              5.44      
Output token throughput (tok/s):         1079.03   
Total Token throughput (tok/s):          2249.77   
---------------Time to First Token----------------
Mean TTFT (ms):                          66936.01  
Median TTFT (ms):                        59901.93  
P99 TTFT (ms):                           169459.47 
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          255.65    
Median TPOT (ms):                        269.80    
P99 TPOT (ms):                           395.92    
---------------Inter-token Latency----------------
Mean ITL (ms):                           220.89    
Median ITL (ms):                         333.07    
P99 ITL (ms):                            427.35    
==================================================
```

### Main (d9ac9e3) chunked prefill off

```
vllm serve ibm-ai-platform/Bamba-9B --port 9999 --no-enable-chunked-prefill --max_model_len=4096
```

```
============ Serving Benchmark Result ============
Successful requests:                     1000      
Benchmark duration (s):                  52.14     
Total input tokens:                      215201    
Total generated tokens:                  198343    
Request throughput (req/s):              19.18     
Output token throughput (tok/s):         3803.80   
Total Token throughput (tok/s):          7930.91   
---------------Time to First Token----------------
Mean TTFT (ms):                          16630.70  
Median TTFT (ms):                        15313.57  
P99 TTFT (ms):                           39832.65  
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          67.94     
Median TPOT (ms):                        62.02     
P99 TPOT (ms):                           257.02    
---------------Inter-token Latency----------------
Mean ITL (ms):                           55.36     
Median ITL (ms):                         56.91     
P99 ITL (ms):                            121.58    
==================================================

## Output Quality

### Bamba-9B

```
vllm (pretrained=ibm-ai-platform/Bamba-9B,tensor_parallel_size=1,dtype=auto,gpu_memory_utilization=0.9,trust_remote_code=True), gen_kwargs: (None), limit: None, num_fewshot: None, batch_size: auto
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.2623|±  |0.0121|
|     |       |strict-match    |     5|exact_match|↑  |0.3700|±  |0.0133|
```

### Zamba2-2.7B

```
vllm (pretrained=Zyphra/Zamba2-2.7B,tensor_parallel_size=1,dtype=auto,gpu_memory_utilization=0.9,trust_remote_code=True), gen_kwargs: (None), limit: None, num_fewshot: None, batch_size: auto
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.5299|±  |0.0137|
|     |       |strict-match    |     5|exact_match|↑  |0.5436|±  |0.0137|
```

### Mamba-Codestral-7B

```
vllm (pretrained=mistralai/Mamba-Codestral-7B-v0.1,tensor_parallel_size=1,dtype=auto,gpu_memory_utilization=0.9,trust_remote_code=True), gen_kwargs: (None), limit: None, num_fewshot: None, batch_size: auto
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.4761|±  |0.0138|
|     |       |strict-match    |     5|exact_match|↑  |0.4632|±  |0.0137|
```