@cyang49 2025-04-25
# Unit tests
```
================================================== test session starts ===================================================
platform linux -- Python 3.12.9, pytest-8.3.4, pluggy-1.5.0
rootdir: /net/storage149/mnt/md0/ccyang/github.com/vllm
configfile: pyproject.toml
plugins: anyio-4.8.0
collected 1570 items                                                                                                     

tests/kernels/mamba/test_causal_conv1d.py ........................................................................ [  4%]
.................................................................................................................. [ 11%]
.................................................................................................................. [ 19%]
.................................................................................................................. [ 26%]
.................................................................................................................. [ 33%]
.................................................................................................................. [ 40%]
.................................................................................................................. [ 48%]
....                                                                                                               [ 48%]
tests/kernels/mamba/test_mamba_mixer2.py ....                                                                      [ 48%]
tests/kernels/mamba/test_mamba_ssm.py ............................................................................ [ 53%]
......................................................................................................s.......s... [ 60%]
.................................................................................................................. [ 68%]
.................................................................................................................. [ 75%]
..............................................                                                                     [ 78%]
tests/kernels/mamba/test_mamba_ssm_ssd.py ........................................................................ [ 82%]
.................................................................................................................. [ 90%]
.................................................................................................................. [ 97%]
..........................................                                                                         [100%]
================================ 1568 passed, 2 skipped, 10 warnings in 881.36s (0:14:41) ================================
```