@cyang49 2025-05-06
@tlrmchlsmth 
Getting a seemingly unrelated test failure. I saw this in other people's PR as well
```
<head></head>
[2025-05-06T03:45:25Z] =========================== short test summary info ============================
--
  | [2025-05-06T03:45:25Z] FAILED entrypoints/openai/test_audio.py::test_chat_streaming_audio[https://vllm-public-assets.s3.us-west-2.amazonaws.com/multimodal_asset/mary_had_lamb.ogg-fixie-ai/ultravox-v0_5-llama-3_2-1b] - AssertionError: assert 'This audio a...t from a poem' == 'This audio a...a traditional'
  | [2025-05-06T03:45:25Z]
  | [2025-05-06T03:45:25Z]   - This audio appears to be a snippet from a traditional
  | [2025-05-06T03:45:25Z]   ?                                           ^^^^^^^ ^^^
  | [2025-05-06T03:45:25Z]   + This audio appears to be a snippet from a poem
  | [2025-05-06T03:45:25Z]   ?                                           ^ ^^
  | [2025-05-06T03:45:25Z] ====== 1 failed, 420 passed, 2 skipped, 36 warnings in 3645.43s (1:00:45) ======
  | [2025-05-06T03:45:28Z] 🚨 Error: The command exited with status 1
  | [2025-05-06T03:45:28Z] user command error: The plugin docker command hook exited with status 1


```